# src/main/java/com/citymate/community/dto/CreateDiscussionRequest.java
package com.citymate.community.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.*;
import java.util.UUID;

@Getter @Setter @NoArgsConstructor @AllArgsConstructor
public class CreateDiscussionRequest {

    @NotNull
    private UUID categoryId;

    @NotBlank
    @Size(min = 3)
    private String title;

    @NotBlank
    @Size(min = 10)
    private String content;
}