# src/main/java/com/citymate/community/dto/CreateReactionRequest.java
package com.citymate.community.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.*;

@Getter @Setter @NoArgsConstructor @AllArgsConstructor
public class CreateReactionRequest {

    @NotBlank
    private String emoji;
}