# src/main/java/com/citymate/community/dto/ErrorResponse.java
package com.citymate.community.dto;

import lombok.*;
import java.time.LocalDateTime;

@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
public class ErrorResponse {
    private LocalDateTime timestamp;
    private int status;
    private String message;
}